% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brq.R
\name{brq}
\alias{brq}
\title{base learner for boosting linear regression quantiles}
\usage{
brq(formula, method = "fn")
}
\arguments{
\item{formula}{a symbolic description of the base learner.}

\item{method}{the algortihm used to fit the quantile regression, the default
is set to "fn", referring to the Frisch-Newton inferior point method.
For more details see the documentation of quantreg::rq.}
}
\value{
brq returns a string, which is used to specifiy the formula in the fitting process.
}
\description{
Base-learner for linear quantile regression.
}
\examples{
brq(cyl * hp)
}
