% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brq.R
\name{brqss}
\alias{brqss}
\title{base learner for boosting nonlinear regression quantiles}
\usage{
brqss(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
base learner for boosting nonlinear regression quantiles
}
