% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.boostrq}
\alias{coef.boostrq}
\title{estimated coefficients of boosting regression quantiles}
\usage{
\method{coef}{boostrq}(object, which = NULL, aggregate = "sum", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
estimated coefficients of boosting regression quantiles
}
