% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.boostrq}
\alias{confint.boostrq}
\title{Test}
\usage{
\method{confint}{boostrq}(
  object,
  parm = NULL,
  level = 0.95,
  B = 1000,
  B.mstop = 25,
  newdata = NULL,
  papply = ifelse(B.mstop == 0, parallel::mclapply, lapply),
  ...
)
}
\arguments{
\item{object}{test}

\item{parm}{test}

\item{level}{test}

\item{B}{tet}

\item{B.mstop}{test}

\item{newdata}{test}

\item{papply}{test}

\item{...}{test}
}
\value{
test
}
\description{
Test
}
\examples{

test <- 1:3

}
