% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv.folds}
\alias{cv.folds}
\title{Folds for k-fold crossvalidation}
\usage{
cv.folds(weights, k = 5)
}
\arguments{
\item{weights}{a numeric vector indicating which weights to used for the single observations.}

\item{k}{number of folds}
}
\value{
a matrix with k columns and n rows, each column indicates
the weights for the respective resampling iteration.
}
\description{
Folds for k-fold crossvalidation
}
\examples{

folds <- cv.folds(rep(1,32), k = 5)
folds

}
