% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cvkrisk.boostrq}
\alias{cvkrisk.boostrq}
\title{k-fold Crossvalidation for boostrq}
\usage{
cvkrisk.boostrq(object, k = 5, grid = 0:mstop(object))
}
\arguments{
\item{object}{a boostrq object}

\item{k}{number of folds, per default 5}

\item{grid}{a vetor of stopping parameters the empirical quantile risk is to be evaluated for.}
}
\value{

}
\description{
k-fold Crossvalidation for boostrq
}
\examples{

}
