% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.boostrq}
\alias{fitted.boostrq}
\title{fitted values of boosting regression quantiles}
\usage{
\method{fitted}{boostrq}(object, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
fitted values of boosting regression quantiles
}
