% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{mstop<-}
\alias{mstop<-}
\title{Updating number of iterations}
\usage{
mstop(x) <- value
}
\arguments{
\item{x}{a boostrq object}

\item{value}{desired number of boosting iterations}
}
\value{
a boostrq object with the updated number of iterations
}
\description{
Updating number of iterations
}
\examples{
model <- boostrq(mpg ~ brq(hp:cyl, cyl*hp) + brq(am), data = mtcars,
mstop = 200, nu = 0.1, tau = 0.5, offset = 0.5)
mstop(model) <- 500
}
