% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{mstop.boostrq}
\alias{mstop.boostrq}
\title{Current number of iterations of boostrq}
\usage{
\method{mstop}{boostrq}(object, ...)
}
\arguments{
\item{object}{a boostrq object}

\item{...}{additional arguments passed to callies}
}
\value{
current number of boosting iterations
}
\description{
Current number of iterations of boostrq
}
\examples{
boosted.rq <-
boostrq(
 formula = mpg ~ brq(cyl * hp) + brq(am + wt),
 data = mtcars,
 mstop = 200,
 nu = 0.1,
 tau = 0.5,
 offset = 0.5
)

mstop(boosted.rq)

}
