% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{mstop.cvkrisk}
\alias{mstop.cvkrisk}
\title{Optimal number of iterations determined by k-fold cross validation}
\usage{
\method{mstop}{cvkrisk}(object, ...)
}
\arguments{
\item{object}{object of class boostrq}

\item{...}{additional arguments passed to callies}
}
\value{
Returns the optimal number of iterations for a boosting regression quantiles model
determined by k-fold cross validation.
}
\description{
Optimal number of iterations determined by k-fold cross validation
}
\examples{
boosted.rq <-
boostrq(
 formula = mpg ~ brq(cyl * hp) + brq(am + wt),
 data = mtcars,
 mstop = 200,
 nu = 0.1,
 tau = 0.5
)

cvk.out <- cvkrisk(boosted.rq, grid = 0:mstop(boosted.rq))

mstop(cvk.out)

boosted.rq[mstop(cvk.out)]

}
