% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{mstop}
\alias{mstop}
\title{s3 method class for 'boostrq'}
\usage{
mstop(object, ...)
}
\arguments{
\item{object}{a boostrq object}

\item{...}{additional arguments passed to callies}
}
\value{
current number of boosting iterations
}
\description{
s3 method class for 'boostrq'
}
\examples{
model <- boostrq(mpg ~ brq(hp:cyl, cyl*hp) + brq(am), data = mtcars,
mstop = 200, nu = 0.1, tau = 0.5, offset = 0.5)
mstop(model)
}
