% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.boostrq}
\alias{print.boostrq}
\title{printing boosting regression quantiles}
\usage{
\method{print}{boostrq}(x, ...)
}
\arguments{
\item{x}{object of class boostrq}

\item{...}{additional arguments passed to callies}
}
\value{
print shows a dense representation of the boostrq model fit.
}
\description{
printing boosting regression quantiles
}
\examples{
boosted.rq <-
boostrq(
 formula = mpg ~ brq(cyl * hp) + brq(am + wt),
 data = mtcars,
 mstop = 200,
 nu = 0.1,
 tau = 0.5,
 offset = 0.5
)

boosted.rq

}
