% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{print.cvkrisk}
\alias{print.cvkrisk}
\title{printing k-fold cross validation results of boosting regression quantiles}
\usage{
\method{print}{cvkrisk}(x, ...)
}
\arguments{
\item{x}{object of class boostrq}

\item{...}{additional arguments passed to callies}
}
\value{
print shows a results of k-fold cross validation of boosting regression quantiles
}
\description{
printing k-fold cross validation results of boosting regression quantiles
}
\examples{
boosted.rq <-
boostrq(
 formula = mpg ~ brq(cyl * hp) + brq(am + wt),
 data = mtcars,
 mstop = 200,
 nu = 0.1,
 tau = 0.5
)

cvk.out <- cvkrisk(boosted.rq, grid = 0:mstop(boosted.rq))

cvk.out

}
