% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.boostrq}
\alias{print.summary.boostrq}
\title{Print result summaries for a boostrq object}
\usage{
\method{print}{summary.boostrq}(x, ...)
}
\arguments{
\item{x}{a summary.boostrq object}

\item{...}{additional arguments passed to callies}
}
\value{
printing the result summaries for a boostrq object including the print-information,
estimated coefficients, and selection frequencies
}
\description{
Print result summaries for a boostrq object
}
\examples{
boosted.rq <-
boostrq(
 formula = mpg ~ brq(cyl * hp) + brq(am + wt),
 data = mtcars,
 mstop = 200,
 nu = 0.1,
 tau = 0.5
)

summary(boosted.rq)

}
