% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.boostrq}
\alias{residuals.boostrq}
\title{residuals of boosting regression quantiles}
\usage{
\method{residuals}{boostrq}(object, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
residuals of boosting regression quantiles
}
