% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{risk.boostrq}
\alias{risk.boostrq}
\title{Empirical Quantile Risk of boostrq Object}
\usage{
\method{risk}{boostrq}(object, ...)
}
\arguments{
\item{object}{a boostrq object}

\item{...}{additional arguments passed to callies}
}
\value{
numeric vector containing the respective empirical quantile risk of the
different boosting iterations.
}
\description{
Empirical Quantile Risk of boostrq Object
}
\examples{
boosted.rq <-
boostrq(
 formula = mpg ~ brq(cyl * hp) + brq(am + wt),
 data = mtcars,
 mstop = 200,
 nu = 0.1,
 tau = 0.5
)

risk(boosted.rq)

}
