% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabsel.R
\name{stabsel.boostrq}
\alias{stabsel.boostrq}
\title{Stability Selection for boosting regression quantiles}
\usage{
\method{stabsel}{boostrq}(
  x,
  cutoff,
  q,
  PFER,
  grid = 0:mstop(x),
  B = ifelse(sampling.type == "MB", 100, 50),
  assumption = "none",
  sampling.type = "MB",
  papply = mclapply,
  FWER,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{test}

\item{cutoff}{test}

\item{q}{test}

\item{PFER}{test}

\item{grid}{stest}

\item{B}{test}

\item{assumption}{test}

\item{sampling.type}{test}

\item{papply}{test}

\item{FWER}{test}

\item{verbose}{test}

\item{...}{test}
}
\value{
test
}
\description{
test
}
\examples{
boosted.rq <-
boostrq(
 formula = mpg ~ brq(cyl * hp) + brq(am + wt),
 data = mtcars,
 mstop = 200,
 nu = 0.1,
 tau = 0.5
)

}
