% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[.boostrq}
\alias{[.boostrq}
\title{Updating number of iterations}
\usage{
\method{[}{boostrq}(x, i, return = TRUE, ...)
}
\arguments{
\item{x}{a boostrq object}

\item{i}{desired number of boosting iterations}

\item{return}{TRUE, if the result should be returned}

\item{...}{additional arguments passed to callies}
}
\value{
a boostrq object with the updated number of iterations
}
\description{
Updating number of iterations
}
\examples{
boosted.rq <-
boostrq(
 formula = mpg ~ brq(cyl * hp) + brq(am + wt),
 data = mtcars,
 mstop = 200,
 nu = 0.1,
 tau = 0.5
)

boosted.rq[500]

}
