% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.boostrq}
\alias{summary.boostrq}
\title{Result summaries for a boostrq object}
\usage{
\method{summary}{boostrq}(object, ...)
}
\arguments{
\item{object}{a boostrq object}

\item{...}{additional arguments passed to callies}
}
\value{
result summaries for a boostrq object including the print-information,
estimated coefficients, and selection frequencies
}
\description{
Result summaries for a boostrq object
}
\examples{
boosted.rq <-
boostrq(
 formula = mpg ~ brq(cyl * hp) + brq(am + wt),
 data = mtcars,
 mstop = 200,
 nu = 0.1,
 tau = 0.5
)

summary(boosted.rq)

}
