% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brq.R
\name{brq}
\alias{brq}
\title{base learner for boosting linear regression quantiles}
\usage{
brq(...)
}
\arguments{
\item{...}{one or more predictor variables.}
}
\value{
brq returns a string, which is used to specifiy the formula in the fitting process.
}
\description{
Base-learner for linear quantile regression.
}
\examples{
brq(hp:cyl, cyl*hp)
}
