% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cvkrisk}
\alias{cvkrisk}
\title{k-fold Crossvalidation for boostrq}
\usage{
cvkrisk(
  object,
  k = 5,
  grid = 0:mstop(object),
  papply = mclapply,
  mc.preschedule = FALSE
)
}
\arguments{
\item{object}{a boostrq object}

\item{k}{number of folds, per default 5}

\item{grid}{a vetor of stopping parameters the empirical quantile risk is to be evaluated for.}

\item{papply}{(parallel) apply function, defaults to mclapply. To run sequentially
(i.e. not in parallel), one can use lapply.}

\item{mc.preschedule}{preschedule tasks if are parallelized using mclapply (default: FALSE)?
For details see mclapply.}
}
\value{
crossvalidation object
}
\description{
k-fold Crossvalidation for boostrq
}
\examples{
boosted.rq <-
boostrq(
 formula = mpg ~ brq(cyl * hp) + brq(am + wt),
 data = mtcars,
 mstop = 200,
 nu = 0.1,
 tau = 0.5
)

set.seed(101)
cvk.out <- cvkrisk(boosted.rq, k = 5, grid = 0:mstop(boosted.rq))

}
