% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.boostrq}
\alias{fitted.boostrq}
\title{fitted values of boosting regression quantiles}
\usage{
\method{fitted}{boostrq}(object, ...)
}
\arguments{
\item{object}{object of class boostrq}

\item{...}{additional arguments passed to callies}
}
\value{
fitted returns the fitted values of the fitted boostrq model.
}
\description{
fitted values of boosting regression quantiles
}
\examples{
boosted.rq <-
boostrq(
 formula = mpg ~ brq(cyl * hp) + brq(am + wt),
 data = mtcars,
 mstop = 200,
 nu = 0.1,
 tau = 0.5,
 offset = 0.5
)

fitted(boosted.rq)

}
