% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.boostrq}
\alias{print.boostrq}
\title{printing boosting regression quantiles}
\usage{
\method{print}{boostrq}(x, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
printing boosting regression quantiles
}
