% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.boostrq}
\alias{residuals.boostrq}
\title{residuals of boosting regression quantiles}
\usage{
\method{residuals}{boostrq}(object, ...)
}
\arguments{
\item{object}{object of class boostrq}

\item{...}{additional arguments passed to callies}
}
\value{
residuals returns the residuals of the fitted boostrq model.
}
\description{
residuals of boosting regression quantiles
}
\examples{
boosted.rq <-
boostrq(
 formula = mpg ~ brq(cyl * hp) + brq(am + wt),
 data = mtcars,
 mstop = 200,
 nu = 0.1,
 tau = 0.5,
 offset = 0.5
)

residuals(boosted.rq)

}
