% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{update.boostrq}
\alias{update.boostrq}
\title{Update and Re-fit a boostrq model}
\usage{
\method{update}{boostrq}(object, index, risk, ...)
}
\arguments{
\item{object}{a boostrq object}

\item{index}{(optional) a index vector indicating which observations should
be used in the fitting process (default: all observations are used).}

\item{risk}{string indicating how the empirical risk should be computed for each boosting iteration.
inbag leads to risks computed for the learning sample (i.e. observations contained in index vector),
oobag to risks based on the out-of-bag (i.e. all observations not in the index vector).}

\item{...}{additional arguments passed to callies}
}
\value{
a re-fitted boostrq model
}
\description{
Update and Re-fit a boostrq model
}
\examples{
boosted.rq <-
boostrq(
 formula = mpg ~ brq(cyl * hp) + brq(am + wt),
 data = mtcars,
 mstop = 200,
 nu = 0.1,
 tau = 0.5
)

update(boosted.rq, index = 1:15, risk = "oobag")

}
